+++ scripts/items.lua
@@ -1,0 +1,18 @@
+-- Passive Digivolution items.
+return function(mod)
+  mod.content.items:register("EGG_OF_KWNLDG", {
+    id = "EGG_OF_KWNLDG",
+    name = "EGG OF KWNLDG.",
+    price = 0,
+    keyItem = true,
+    tossable = false,
+  })
+
+    mod.content.items:register("EGG_OF_COURAGE", {
+    id = "EGG_OF_COURAGE",
+    name = "EGG OF COURAGE",
+    price = 0,
+    keyItem = true,
+    tossable = false,
+  })
+end